worker/requirements.txt
whisperx==3.7.6
celery
redis
fleep
pympi-ling
yt-dlp
pydub
ffmpeg-normalize
SciPy
lingua-language-detector